2022x.b001. Fixed bug in export of deprecated QuantityKind with temp try-catch block.

+++ src/main/java/org/aero/mtip/metamodel/core/CommonElement.java
@@ -394,7 +394,7 @@ public abstract class CommonElement {
 						.collect(Collectors.toList());
 		} else if (valueType.contentEquals(XmlTagConstants.TV_TYPE_REAL)) {
 				return values.stream()
-						.map(x -> Float.toString((Float)x))
+						.map(x -> Double.toString((Double)x))
 						.collect(Collectors.toList());
 		} else if (valueType.contentEquals(XmlTagConstants.TV_TYPE_STRING)) {
 				return values.stream()
